bug fixes and update unit tests
