+++ src/applications/vre/28-8832/sass/28-8832-planning-and-career-guidance.scss
@@ -6,3 +6,9 @@
 @import "~@department-of-veterans-affairs/formation/sass/modules/m-modal";
 @import "~@department-of-veterans-affairs/formation/sass/modules/m-omb-info";
 @import "../../../../platform/forms/sass/m-form-confirmation";
+
+.usa-alert {
+    &:before {
+      display: none;
+    }
+  }
